% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_ts_sim.R
\name{arima_ts_sim}
\alias{arima_ts_sim}
\title{Simulates time series using univariate ARIMA models}
\usage{
arima_ts_sim(
  intercept = 0,
  noise_mean = 0,
  noise_sd = 1,
  x1_noise_mean = 0,
  x1_noise_sd = 1,
  effect = 0,
  effect_ceiling = Inf,
  ts_length = 200,
  treat_start = 75,
  coefficient_x1 = 1,
  change_type = "stasis",
  delta = 1,
  round_to_integers = FALSE,
  seed = 175,
  ...
)
}
\arguments{
\item{intercept}{The starting y-intercept value for the time series.}

\item{noise_mean}{The average value of random noise that should be added to the time series.}

\item{noise_sd}{The standard deviation of random noise that should be added to the time series.}

\item{x1_noise_mean}{The average value of to random noise that should be added when simulating x1.}

\item{x1_noise_sd}{The standard deviation of random noise that should be added when simulating x1.}

\item{effect}{The initial effect of the intervention.}

\item{effect_ceiling}{The maximum value for the pointwise effect of a policy. This argument only works when \code{change_type = "amplification"}.}

\item{ts_length}{The length of the simulated time period.}

\item{treat_start}{The time that the intervention begins.}

\item{coefficient_x1}{Scales the value of x1.}

\item{change_type}{Specifies the way that the treatment should change over time. Possible inputs are \code{stasis}, \code{attenuation}, and \code{amplification}.}

\item{delta}{Specifies that rate that the effect of the treatment changes.}

\item{round_to_integers}{Should the simulated values be rounded to integers?}

\item{seed}{Sets the random seed.}
}
\description{
This function simulates time series and allows users to add interventions
whose effects vary over time. The function returns a tibble.
}
