% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsim_effect_plot.R
\name{dsim_effect_plot}
\alias{dsim_effect_plot}
\title{Simulates time series using univariate ARIMA models}
\usage{
dsim_effect_plot(dsim_object, plot_output = "treated_time_series")
}
\arguments{
\item{dsim_object}{The output from \code{arima_ts_sim}.}

\item{plot_output}{Specifies which time series should be plotted on the y-axis. There are four possible options: \code{pointwise_effect}, \code{cumulative_effect}, \code{untreated_time_series}, and \code{treated_time_series}.}
}
\description{
This function plots the  simulated time series in \code{ggplot2}.
}
