% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsim_effect_plot.R
\name{dynamic_effect_gg_theme}
\alias{dynamic_effect_gg_theme}
\title{Simulates time series using univariate ARIMA models}
\usage{
dynamic_effect_gg_theme(base_size = 14)
}
\arguments{
\item{dsim_object}{The output from \code{arima_ts_sim}.}

\item{plot_output}{Specifies which time series should be plotted on the y-axis. There are four possible options: \code{pointwise_effect}, \code{cumulative_effect}, \code{untreated_time_series}, and \code{treated_time_series}.}
}
\description{
This function plots the  simulated time series in \code{ggplot2}.
}
