% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_n_consec_rows.R
\name{group_n_consec_rows}
\alias{group_n_consec_rows}
\title{Create group IDs for specified number of consecutive rows in data frames}
\usage{
group_n_consec_rows(data, group.size = 7)
}
\arguments{
\item{data}{Data frame}

\item{group.size}{Number of consecutive rows to group together}
}
\description{
Create group IDs for specified number of consecutive rows in data frames
}
