% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_sim_data.R
\name{aggregate_sim_data}
\alias{aggregate_sim_data}
\title{Aggregate simmulated time series data by group}
\usage{
aggregate_sim_data(
  data,
  agg_function = "sum",
  window_size = 7,
  drop_incomplete_window = TRUE
)
}
\arguments{
\item{data}{Data frame}

\item{agg_function}{#' @param window_size The number of consecutive time intervals to group together
#' @param drop_incomplete_window}
}
\description{
Aggregate simmulated time series data by group
}
